GEODE-7613: move multicast address init to where it's needed
